Manually upgrade the versions
I tried the recommended `npx sb@latest upgrade` but the stories wouldn't
load properly after this. Doing things incrementally to try to identify
the problem.

I first confirmed that `6.1.15` works, and now that 21 works I'm
comitting it. The html addon seems to be the only thing that doesn't
work, so I'll try to fix that next