Change container width based on viewport
I'm doing this to show that the component respects the width. On mobile
the content container will likely be full width, so the button will be
against the right of the screen. On larger screens the content will
likely be in a column layout with whitespace on the side, and in that
case the button will only go as right as its parent content container,
not to the full right side of the screen