Add Back to top Web Component (#170)
* Add initial behavior of going back to top

The styling is very rough, and other behaviors haven't been
tested/implemented.

* Add behavior for "revealing" the button

* Add docking behavior

Refactoring needed :sweat-smile:

* Change the docking styles

This allows us to just apply the style to the parent div. It also
anchors the button to the right of the dock

* Use opacity properly

* Right-align button to parent when revealed

This applies when the button isn't docked

* Copy styles from usa-button class

* Set the margin

* Add icon & some styles for responsiveness

* Remove icon class & wrapping span

Just trying to reduce the amount of markup

* Add hover state

* Add an animation

* Use IntersectionObserver for docking behavior

* Remove unused param

* Use IntersectionObserver for reveal behavior

* Remove isInViewport function

The IntersectionObserver replaces this

* Remove some redundant CSS

* Simplify the docking callback

* Inline the observer callbacks

* Switch all observers to componentDidLoad

This is just to make things more consistent. The observer for
`this.revealPixel` needs to make the observe call after the render so it
can be initialized.

* Use classnames dependency

* Use a Prop for the breakpoint value

* Add animate-css dependency

I don't like the relative import from node_modules, but I couldn't get
stencil/postcss-import to work like I wanted it to.

* Add test for scrolling reveal

* Add test for docking behavior

* Add test to check for onClick behavior

* Tweak styles a bit

* Move some styles around

* Remove some class properties

We no longer need them to be defined at the class level now that they
are all used in the same function

* Remove breakpoint Prop

I don't think we need a breakpoint prop when this can just be changed with
the CSS variable `--reveal-breakpoint`

* Make sure button stays docked on huge footers

* Add test to cover previous commit

* Add basic axe check

* Add axe check for other behaviors

* Be more granular with the skip/ignore

* Add some comments based on review feedback

* Remove unused CSS

* Remove one more unused CSS & swap px for rem

* Rework all the tests

Now they all share the same page structure, and the scrolling heights
have been given names to indicate the reason for the distance

* Be more specific in axeCheck tests

* Bump dependency versions

Doing this seems to reduce some flakiness

* Fix last flaky test :crossed-fingers:

* Switch name of state variable

* Don't remove right border radius on desktop

It's being removed by default for mobile view, so this adds it back in

* Update doc comment

Co-authored-by: Chris Valarida <chris.valarida@adhocteam.us>

* Respect user preferences for reduced motion

* Ensure the button is announced as a button

This is from a11y feedback:
https://github.com/department-of-veterans-affairs/va.gov-team/issues/31217

* Rework CSS to enable tabbing to button when hidden

This is based on a11y feedback:
https://github.com/department-of-veterans-affairs/va.gov-team/issues/31218

* Update failing tests

* v0.12.0

Co-authored-by: Chris Valarida <chris.valarida@adhocteam.us>