Add back-to-top component to Storybook (#171)
* Switch to btt branch

* Add initial story for BTT web component

This is to get the chromatic deployment so I can do manual testing in
IE11.

* Add IE11 polyfills to Storybook

I'm doing this so that I can use our Chromatic deployments to test
components from IE11, rather than going through the whole process of
running a local server on my machine with ports exposed so that I can
access it from my Windows VM

* Get latest from branch

* Get version w/ IntersecionObserver

* Update yarn.lock

* Bump to latest development branch

* Change container width based on viewport

I'm doing this to show that the component respects the width. On mobile
the content container will likely be full width, so the button will be
against the right of the screen. On larger screens the content will
likely be in a column layout with whitespace on the side, and in that
case the button will only go as right as its parent content container,
not to the full right side of the screen

* Mark the old component as deprecated

* Remove pasted commented out code

* Clean up the story

* Bump version

* Use the component release

* Change footer background color for story

* Make footer text white