Integrate web-components branch, Pt. 3 (#199)
* Move web-components prettier config to root

* Prettify the core webpack config

* Use webpack plugin to copy build files

This is meant to allow for a gentle transition to the new named import
strategy that `core` is providing.

* Try a yarn install at the root

* Create new branch to test all the CI changes (#200)

* Create new branch to test all the CI changes

This is to keep the commit history from getting cluttered in the main PR

* See if a specific package-level install helps

* Do the same install in chromatic job

* Try installing dependencies in storybook package

* Just try building storybook directly

* Build core package

* Try publishing using action & pre-built storybook

* Use package name instead of directory :facepalm:

* Change branch back to master

* Test deploying storybook to dev environment

This commit is based on the changes to the chromatic job in #200

* Include current branch for testing

* Get publish-storybook into ready state

Undo the testing changes

* Move the copied files up out of dist

* Include dist dir and JS files in published package

* Remove storybook config from web-components dir

* See if Date prop descriptions show up

* Import directly from source file instead of index

* Use a webpack alias to use React component source

This will give us the propTypes and associated information for each
component. It gets stripped away on build, which means the args table
for React component stories are barren.

* Revert back to the json file import for WC docs

Aliasing the component-library to react-components in storybook allows
us to use the React source with propTypes, but we also need to account
for importing web-components items from component-library

* Do some cleanup

* Add some react dependencies to core

I discovered that we needed this in order for things to build without
errors in vets-website. This is because react-components doesn't have a
proper bundle, so its dependencies aren't included.

* Include ProgressButton in index files

* Move react-components README to core

* Add stub README at root level

* Add comment from review feedback

* Copy react-bindings into core build output

During some testing I found that importing a twice-built React bindings
from `component-library` lead to errors. The binding had been turned
into a Class, while the original binding from `web-components` was an
Object. I'm not sure why webpack did this, but since we're already using
the `webpack-copy-plugin` anyway I decided to copy over the
`react-bindings` that I knew would work.