Add functionality for analytics event
There's a bit of awkwardness here. I'm not doing the traditional
`this.componentLibraryAnalytics.emit(...)`, because we want to fire the
event once the loading indicator is gone. Doing that in
`disconnectedCallback` is too late - the node that would fire the event
has already been removed from the DOM. Using the mutation observer
allows us to fire the event from the parent instead, which isn't ideal.