Add loading indicator web component (#207)
* Initial draft for loading indicator component

Still need to add styling, focus behavior, & analytics

* Add story for loading indicator component

* Copy over CSS from formation

* Add focus behavior

* Add axeCheck test

* Change indicator size to 4rem

* Add functionality for analytics event

There's a bit of awkwardness here. I'm not doing the traditional
`this.componentLibraryAnalytics.emit(...)`, because we want to fire the
event once the loading indicator is gone. Doing that in
`disconnectedCallback` is too late - the node that would fire the event
has already been removed from the DOM. Using the mutation observer
allows us to fire the event from the parent instead, which isn't ideal.

* Cast to an appropriate type

* Update story for analytics event

* Use mutation record to determine removal

* Ensure the mutation observer callback runs

* Update the base test

* Clean up stories a bit

* Tweak the size of the spinner

* Remove commented code

* Mark old LoadingIndicator as deprecated

* Set tabindex to -1

This is from staging review feedback:
https://github.com/department-of-veterans-affairs/va.gov-team/issues/32577

* Slow spinner down & make it smaller

In response to staging review feedback:
https://github.com/department-of-veterans-affairs/va.gov-team/issues/32583

* Add aria-live="polite"

Staging review feedback:
https://github.com/department-of-veterans-affairs/va.gov-team/issues/32813

* Fix test

* Bump versions to prepare for release & publish