Add additional info web component (#221)
* Add rought draft for additional-info component

* Only show border when open

* Add icon

* Add keyboard control & other things

Where "other things" is copying over some of the markup structure from
the React component.

* Add analytics event

* Add story

* Add axeCheck tests

* Update basic render test

* Update story contents

* Tweak styles for parity

* Add test for keyboard control

* Include test for click expansion

* Add context to the story

* Put listeners directly on the anchor tag

This is to prevent clicks in the rest of the component from toggling the
expansion/collapse

* Add another analytics test

* Remove offset by setting negative margin

This means that we have the border issue documented here:
https://github.com/department-of-veterans-affairs/vets-design-system-documentation/issues/577

* Mark AdditionalInfo as deprecated

* Remove copypasta'd !important

* Move blue border down a level

To address the issue with it overlapping/not appearing:
https://github.com/department-of-veterans-affairs/vets-design-system-documentation/issues/577

* Move aria-expanded down to anchor

This will hopefully announce the correct state to screen readers

* Let Enter key trigger expansion

* Fix styles after aria-expanded change

* Hide slotted content when closed

Setting opacity doesn't work for screen readers, and the display prop
cannot be animated. This should hide the slotted content from screen
readers.

https://developer.mozilla.org/en-US/docs/Web/CSS/visibility#accessibility_concerns

* Unify colors after meeting with Natalie

This was discussed in the council. The color that was originally
suggested wasn't in the design system, so we've chosen a pre-existing
color that's just slightly different.

* Bump minor version for publishing