Override valueOf on row for generic comparison
This may not be the option that we go with, but I wanted to put it in
the git history anyway. This allows the `partition()` function to use
simple relational operators without the need to complicate it by
selecting which column in the row needs to be used for a comparison.
However, overriding the default `valueOf()` method to return only part
of the content just to achieve simpler sorting might not be great.

If we went with this option, `<va-table>` would set an appropriate prop
on its slotted `<va-table-row>` row elements so that `valueOf()` could
return the right cell content.