Remove the middleman i and j variables
Modifying the integer parameters is technically safe since we're
changing values, not references. There's still a part of me that is a
bit nervous about it though. Still, I'm doing it because I think it's
more clear than having parameters like `loStart` and `hiStart` which we
then assign to `lo` and `hi` for additional modification.