Add sorting functionality to va-table component (#318)
* Commit rough implementation of quicksort

* Change quicksort implementations

* Update this.rows after sort finishes

This is crucial and ensures that any re-sorts start out with an up to
date row list. Data will get scrambled otherwise

* Override valueOf on row for generic comparison

This may not be the option that we go with, but I wanted to put it in
the git history anyway. This allows the `partition()` function to use
simple relational operators without the need to complicate it by
selecting which column in the row needs to be used for a comparison.
However, overriding the default `valueOf()` method to return only part
of the content just to achieve simpler sorting might not be great.

If we went with this option, `<va-table>` would set an appropriate prop
on its slotted `<va-table-row>` row elements so that `valueOf()` could
return the right cell content.

* Change swap function to use indices as args

* Fix index bug

* Go back to a selector function

This is still cumbersome due to having to pass the function around a
lot. I may look at other options

* Put a button inside the sortable column

* Indicate empty table cell with CSS instead of JS

* Import SVG & inline it in the DOM

* Move sorting function to button click handler

This also tweaks some CSS - will clean it up later

* Pull quicksort into separate file

I will eventually add unit tests & fix any bugs

* Refactor quicksort to work with NodeList

In my unit tests I had a bug because the `rows` argument was an array
which wasn't being updated.

There's also a bugfix for making sure we don't skip the pivot when
sorting.

* Let sortColumn be any index

* Add story dedicated to sorting

* Add reverse sort functionality

Still need to add tests

* Change default selector to return textContent

This should be more straightforward since we know that we will be
comparing strings instead of Node/Element objects. Documentation and
tests to come.

* Move slotted CSS into separate file

Trying to help with maintenance

* Use the right path

* Add styles to match curent react implementation

* Adjust icon naming & aria-label

* Allow initial sort state to be descending

By ensuring that sortable tables are initially sorted ascending or
descending, we should be able to ensure that the aria-label of the
button is correct. Allowing for an initial "unsorted" state, then
proceeding to sorting by ascending/descending would complicate things.

* Commit progress so far on testing sort in table

I haven't been able to successfully test the table after its children
have been sorted. I also tried adding a unit test for this instead of
e2e, but jest is having trouble with the svg import, particularly the
`?format=text` part.

* Add test for button on descending sort

* Add comments about missing tests

* Add basic unit tests for reverseQuicksort

* Use the default selector

* Add 2 more unit tests for sorting w/ numbers

* Reorder some statements

* Add comments for quicksort helper

* Refactor swapChildren & add comments

* Use a more descriptive variable name for the pivot

* Fix broken tests

The tests were only failing after the merge - I'm not sure what went
wrong.

* Remove the middleman i and j variables

Modifying the integer parameters is technically safe since we're
changing values, not references. There's still a part of me that is a
bit nervous about it though. Still, I'm doing it because I think it's
more clear than having parameters like `loStart` and `hiStart` which we
then assign to `lo` and `hi` for additional modification.

* Don't truncate variable names

This was leftover from copying some quicksort pseudocode

* Rename util file from dom to dom-sort

* Update to better handle plain HTML children

childNodes includes text nodes, which is a problem in plain HTML files
where the indentation gets counted as a node. `children` fixes that, but
this still doesn't work in IE :/

* Address QA feedback for changing CSS

The ticket: https://github.com/department-of-veterans-affairs/va.gov-team/issues/39405

* Add aria-sort attribute on column header

This is the feedback requested in:

- https://github.com/department-of-veterans-affairs/va.gov-team/issues/39433
- https://github.com/department-of-veterans-affairs/component-library/issues/334#issuecomment-1083663371
- https://github.com/department-of-veterans-affairs/va.gov-team/issues/36938#issuecomment-1084842892

* Remove warning in story