Add infrastructure for internationalization (#346)
* Add initial setup for using i18n in web-components

* Setup i18n demo for a React component

This uses shared config. There's some build changes since the
react-components build is awkward and doesn't use a real bundler, so we
need to move things in where they'll be found

* Fix lock file

* Add the actual setup file

* Add the other helper files

* Temporarily change storybook's build settings

This is to show that importing from the root-level component library
will result in web components having i18n support

* Try to get rendered components to update

I'm trying to have this happen automatically when the lang attribute
changes

* Be specific about the language to change to

* Start with english instead

* Remove one of the imports

* Move some files around for better grouping

* Fix the filenames glob

We want to make sure the translation files are moved as well

* Move languageChanged listener into callback

This should make a bit more sense since we only want to register this
handler once - we don't need to do it before each render.

* Change state when a language change is detected

The state change will trigger a re-render. I'll see if there is a better
way of accomplishing this

* Add i18next-browser-languagedetector to peer deps

* Try to fix e2e tests

I wasn't able to get the `cimode` working if I set the language in the
test file, but doing it in the actual source file worked. I think it's
because that gets bundled up, and the `i18next` instance that gets
imported in the test file is different from the bundled instance in the
component source, so the `.init()` is working off of different instances

* Use main tag for lang detection, not content id

The `<main>` tag is already important as an accessibility landmark and
_should_ be on every page:

https://dequeuniversity.com/rules/axe/4.4/landmark-one-main

Any random element with `id="content"` doesn't really matter, so by
targeting the `<main>` element we are saying that our components can
dynamically switch language, but we're going to look for standard
elements for a language indicator.

* Actually change the language detector :facepalm:

* Add a demo of translation w/ interpolation

This also fixes a bug - if value is null, value.length would throw an
error

* Prepare versions for prerelease

This is to enable the check-in team to do some testing with a few
limited components.

* Remove comments in mutation observer

* Only set up observer if main element exists

* Update test for max-chars translation

* Undo pre-release version changes

* Undo changes to SimpleDate React component

* Undo storybook Date changes

* Undo storybook import changes

* Remove changes to va-text-input

* Undo changes to va-text-input stories

* Remove translations

* Apply prettier formatting

* Leave the translation imports as-is