Try to set the MutationObserver up faster
I think the `load` event handler was added for testing in storybook.
This slows down the observer setup which means an app may modify
`<main>` before our observer is looking for it. I'm hoping this
short-circuit loop will allow the observer to be setup as soon as
possible while also allowing for storybook to render the `<main>` tag
through React which will take some extra milliseconds