Rework date validation (#506)
* Commit rough draft of validation changes

* Reflect error prop into DOM directly

This is the more idiomatic way and removes less clutter from the JSX in
the render function

* Reflect the value prop in the same way

I don't know why this was reflected originally, but it's better to do
that in the prop options rather than in the JSX

* Update some E2E tests

This is just to make a bit of progress - the tests may end up being
changed around a lot

* Move validation into util file

The goal is to get this centralized so we have the same validation
running on va-date and va-memorable-date. It looks like passing the
component by reference still allows us to set props

* Reuse validate function in memorable-date

This has some other refactoring as well. I'd like to come up with a
better solution for handling monthYearOnly.

* Export checkLeapYear function for unit test

* Update unit tests to use correct type

* Adjust some tests & add day validation

This is kind of messy because of all the tests I'm dealing with. Will
try to clean up a bit later.

* Adjust day validation

* Group all Date props together

* Update some more tests

* Mark error & value as reflected & mutable

* Make slight refactoring to some date tests

* Regroup some memorable date tests

* Fix bug in day validation

* Add some monthYearOnly tests

* Add test for month validation to memorable-date

* Add another test for day validation

* Add more detail to invalid month test

* Add some comments

* Add some missing tests

* Add tests for invalid* props

* Remove apparently unneeded change

* Invalidate year in custom validation demo

* Remove bug where day is invalidated unnecessarily

I think this was added as I was trying to get some tests to pass, but I
didn't test it in storybook.

* Bump version to prepare for publishing